+++ lib/pushmi_pullyu/swift_depositer.rb
@@ -1,9 +1,6 @@
 require 'digest/md5'
 require 'openstack'
-
 class PushmiPullyu::SwiftDepositer
-
-  attr_reader :swift_connection
 
   def initialize(connection)
     # Generic authentication parameters
@@ -13,7 +10,11 @@ class PushmiPullyu::SwiftDepositer
       auth_url: connection[:auth_url],
       project_name: connection[:project_name],
       auth_method: 'password',
-      service_type: 'object-store'
+      service_type: 'object-store',
+      # The retry_auth value should be true by default. It is currently set
+      # to nil which causes the connection to die without giving more error
+      # details.
+      retry_auth: true
     }
 
     if connection[:auth_version] == 'v3'
@@ -28,10 +29,8 @@ class PushmiPullyu::SwiftDepositer
 
   def deposit_file(file_name, swift_container)
     file_base_name = File.basename(file_name, '.*')
-
     checksum = Digest::MD5.file(file_name).hexdigest
-
-    era_container = swift_connection.container(swift_container)
+    era_container = @swift_connection.container(swift_container)
 
     # Add swift metadata with in accordance to AIP spec:
     # https://docs.google.com/document/d/154BqhDPAdGW-I9enrqLpBYbhkF9exX9lV3kMaijuwPg/edit#